pyproject.toml
[project]
name = "rag-personalized-agentic-tutor"
version = "0.1.0"
description = "A personalized AI tutor built as a multi-agent RAG system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "streamlit>=1.42.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "loguru>=0.7.0",
    "docling>=2.5.0",
    "pymupdf>=1.25.0",
    "surya-ocr>=0.6.0",
    "langchain-text-splitters>=0.3.0",
    "google-generativeai>=0.8.0",
    "fastembed>=0.4.0",
    "qdrant-client>=1.11.0",
    "redis>=5.0.0",
    "tenacity>=9.0.0",
    "tiktoken>=0.7.0",
    "Authlib>=1.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ipykernel>=6.0",
    "fpdf2>=2.8.0",
]
gpu = [
    "colpali-engine>=0.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
